+++ yiilian-dht/src/data/get_peers_reply.rs
@@ -30,7 +30,7 @@ pub struct GetPeersReply {
     pub ip: Option<SocketAddr>,
 
     /// readonly
-    pub ro: Option<i32>,
+    pub ro: Option<u8>,
 
     // ----------------------------
     /// sender node id
@@ -55,7 +55,7 @@ impl GetPeersReply {
         t: TransactionId,
         v: Option<Bytes>,
         ip: Option<SocketAddr>,
-        ro: Option<i32>,
+        ro: Option<u8>,
     ) -> Self {
         Self {
             id,
